+++ js_libraries/lynx-core/src/lynx/interface.ts
@@ -49,8 +49,6 @@ export interface NativeLynxProxy extends BackgroundLynx {
   stopRecording(value: object): void;
 
   fetch: (input: RequestInfo, init?: RequestInit) => Promise<Response>;
-
-  loadScript(url: string, options?: { bundleName?: string }): object;
 }
 
 export const enum ContextProxyType {